+++ packages/noob-core/src/tests/toposort.rs
@@ -38,7 +38,7 @@ fn static_order_with_groups(interner: &mut Interner, sorter: &mut Sorter) -> Vec
     let mut groups = Vec::new();
     while sorter.is_active() {
         let ready = sorter.get_ready(interner, None);
-        let out: Vec<u32> = sorter.out.iter().copied().collect();
+        let out: Vec<u16> = sorter.out.iter().copied().collect();
         if ready.is_empty() && out.is_empty() {
             // python's generator would loop forever here; fail loudly instead
             panic!("sorter is_active() but nothing is ready or out");
@@ -62,7 +62,7 @@ fn test_graph(graph: &[(&str, &[&str])], expected: &[&[&str]]) {
     let mut sorter = Sorter::default();
     for (node, deps) in graph {
         let node_id = interner.intern_node(node);
-        let dep_ids: Vec<u32> = deps.iter().map(|d| interner.intern_node(d)).collect();
+        let dep_ids: Vec<u16> = deps.iter().map(|d| interner.intern_node(d)).collect();
         sorter.add(&mut interner, node_id, &dep_ids, true).unwrap();
     }
 
@@ -193,7 +193,7 @@ fn assert_cycle(graph: &[(&str, &[&str])], cycle: &[&str]) {
     let mut sorter = Sorter::default();
     for (node, deps) in graph {
         let node_id = interner.intern_node(node);
-        let dep_ids: Vec<u32> = deps.iter().map(|d| interner.intern_node(d)).collect();
+        let dep_ids: Vec<u16> = deps.iter().map(|d| interner.intern_node(d)).collect();
         sorter.add(&mut interner, node_id, &dep_ids, true).unwrap();
     }
     let found: Option<Vec<String>> = sorter.find_cycle().map(|c| {
@@ -359,7 +359,7 @@ fn test_optional_dependencies() {
 
     sorter.mark_expired(&[a_a1], true);
     sorter.done(&interner, &[a_a2]).unwrap();
-    let ready: IndexSet<u32> = sorter.get_ready(&interner, None).into_iter().collect();
+    let ready: IndexSet<u16> = sorter.get_ready(&interner, None).into_iter().collect();
     assert_eq!(ready, IndexSet::from([only_optional, mixed, b]));
 
     sorter.done(&interner, &[only_optional, mixed]).unwrap();
@@ -378,9 +378,9 @@ fn unlock_optionals_case(unlock_optionals: bool) {
     let ready = sorter.get_ready(&interner, None);
     sorter.done(&interner, &ready).unwrap();
 
-    let out: Vec<u32> = sorter.out.iter().copied().collect();
+    let out: Vec<u16> = sorter.out.iter().copied().collect();
     sorter.mark_expired(&out, unlock_optionals);
-    let ready: IndexSet<u32> = sorter.get_ready(&interner, None).into_iter().collect();
+    let ready: IndexSet<u16> = sorter.get_ready(&interner, None).into_iter().collect();
     if unlock_optionals {
         let expected = IndexSet::from([
             interner.intern_node("only_optional"),
